feat: 商品加购接口 60%

+++ main/app/model/product.go
@@ -79,7 +79,7 @@ type ProductPackageAttribute struct {
 type ProductPackage struct {
 	BaseModel
 	Name                  string `gorm:"default:'';column:name;comment:'产品包名称'"`
-	MultiLanguageNameUuid uint   `gorm:"default:0;column:multi_language_name_uuid;comment:'多语言名称UUID'"`
+	MultiLanguageNameUuid uint64 `gorm:"default:0;column:multi_language_name_uuid;comment:'多语言名称UUID'"`
 	ImageName             string `gorm:"default:'';column:image_name;comment:'图片名称'"`
 	ImageFileUuid         uint64 `gorm:"default:0;column:image_file_uuid;comment:'图片UUID'"`
 	DeductStockType       uint   `gorm:"default:0;column:deduct_stock_type;comment:'库存计算方法, 0-下单减库存 1-付款减库存'"`
